Winter 2024/WS24-Software-Project/text-simplification-main/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
pytest==8.3.3
flask-testing==0.8.1
pytest-flask==1.3.0
coverage==7.6.4
pytest-cov==6.0.0
spaCy==3.8.2
pygls==1.0.0
transformers==4.28.1
sentencepiece
torch
flask-restx
